Energy map is now generated using mof path instead of MOF object.
energy_map method is modified to take mof_path as argument.
An instance of MOF object is created inside the energy_map method.
This requires force_field information which is now passed as an argument to energy_map method.

Also now the structure extension is performed in energy_map method.