Number of structures for reporting and exporting are now separated.
The number of structures to report in results.yaml and the number of
structures to export cif files for can be controlled separately
from the simulation parameters file (sim_par.yaml).

Of course both numbers are adjusted according to the actual number of
structures discovered in interpenetration by taking the minimum btw
requested number and structures discovered.

Also, fixed a bug in regenerate_structures (min instead of max).